Updating CarInfoActivity
Updates made the the CarInfoActivity include the addition of spinners to
the selection page, and the integration of vehicle selection from
FuelEconomy.gov. Changes to the Vehicle class include changes to the
Vehicle() method. Important: Forced Portrait orientation on the
CarInfoActivity window.

Additions: New DataParser class that once refactoring is complete will
contain the AsyncDownloader classes from the CarInfoActivity class.

Fixes still required: Clean up the timing issues on the CarInfoActivity.
Connect the output to the MainActivity. Use Vehicle ID to request MPG
for Vehicle. General code cleaning.